backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
passlib[bcrypt]
python-jose[cryptography]
dotenv
python-multipart